+++ src/field_door.c
@@ -224,7 +224,7 @@ bool8 FieldIsDoorAnimationRunning(void)
     return FuncIsActiveTask(Task_AnimateDoor);
 }
 
-u8 sub_8058790(u32 x, u32 y)
+u32 sub_8058790(u32 x, u32 y)
 {
     if (cur_mapdata_get_door_x2_at(gDoorAnimGraphicsTable, x, y) == 0)
         return 8;
